Threshold + fix mid-line start
Increased the movement threshold to 5.
Also now it waits until it's done a swing back (wait_for_mouse(-5)) and is going in the positive direction again (wait_for_mouse(5)) before it starts drawing. This should fix the case where it finishes a line in the middle and immediately starts on the next one.